+++ CHANGELOG.md
@@ -12,11 +12,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add more tests.
 - Add `JavaDoc`.
 
-## [1.17.3](https://github.com/appulse-projects/utils-java/releases/tag/1.17.3) - 2020-02-25
+## [1.18.0](https://github.com/appulse-projects/utils-java/releases/tag/1.18.0) - 2020-02-25
 
 ### Added
 
-- A **read-only** implementation of `Bytes`.
+- A **read-only** implementation of `Bytes`;
+- Now it's possible to create a diffirent `slice`s of `Bytes`.
 
 ### Changed